Partitioner: global_index_by_pid_map is constructed in
partitioner instead of parmetis_paritioner

becuase other partitioner such petsc_partitioner also need this.
