+++ include/partitioning/partitioner.h
@@ -28,6 +28,7 @@
 // C++ Includes
 #include <cstddef>
 #include <memory>
+#include <unordered_map>
 
 namespace libMesh
 {
@@ -211,9 +212,29 @@ protected:
   static const dof_id_type communication_blocksize;
 
   /**
+   * Construct contiguous global indices for the current partitioning. The global indices
+   * are ordered part-by-part
+   */
+   virtual void _find_global_index_by_pid_map(const MeshBase & mesh);
+
+  /**
    * The weights that might be used for partitioning.
    */
   ErrorVector * _weights;
+
+  /**
+   * Maps active element ids into a contiguous range, as needed by parallel partitioner.
+   */
+  std::unordered_map<dof_id_type, dof_id_type> _global_index_by_pid_map;
+
+  /**
+   * The number of active elements on each processor.
+   *
+   * \note ParMETIS requires that each processor have some active
+   * elements; it will abort if any processor passes a NULL _part
+   * array.
+   */
+  std::vector<dof_id_type> _n_active_elem_on_proc;
 };
 
 } // namespace libMesh
